requirements.txt
discord.py==1.5.1
psutil==5.7.3
psycopg2==2.8.6
jikanpy==4.2.2
dblpy==0.4.0
git+git://github.com/Rapptz/discord-ext-menus
